split out non-HTTP server parts
